adding fix for http payload, adding injection of test step action if unhandled error thrown
