+++ testah-junit/src/main/java/org/testah/driver/http/requests/PatchRequestDto.java
@@ -2,6 +2,7 @@ package org.testah.driver.http.requests;
 
 import org.apache.http.HttpEntity;
 import org.apache.http.client.methods.HttpPatch;
+import org.testah.TS;
 
 public class PatchRequestDto extends AbstractRequestDto {
 
@@ -22,6 +23,20 @@ public class PatchRequestDto extends AbstractRequestDto {
     public PatchRequestDto(final String uri, final HttpEntity payload) {
         super(new HttpPatch(uri), "PATCH");
         setPayload(payload);
+    }
+
+    public AbstractRequestDto setPayload(final HttpEntity payload) {
+        try {
+            if (null != payload) {
+                httpEntity = payload;
+                ((HttpPatch) httpRequestBase).setEntity(payload);
+            } else {
+                TS.log().warn("payload was null so ignoring");
+            }
+        } catch (final Exception e) {
+            throw new RuntimeException(e);
+        }
+        return this;
     }
 
 }
