+++ testah-junit/src/main/java/org/testah/driver/http/requests/PostRequestDto.java
@@ -2,6 +2,7 @@ package org.testah.driver.http.requests;
 
 import org.apache.http.HttpEntity;
 import org.apache.http.client.methods.HttpPost;
+import org.testah.TS;
 
 public class PostRequestDto extends AbstractRequestDto {
 
@@ -22,6 +23,20 @@ public class PostRequestDto extends AbstractRequestDto {
     public PostRequestDto(final String uri, final HttpEntity payload) {
         super(new HttpPost(uri), "POST");
         setPayload(payload);
+    }
+
+    public AbstractRequestDto setPayload(final HttpEntity payload) {
+        try {
+            if (null != payload) {
+                httpEntity = payload;
+                ((HttpPost) httpRequestBase).setEntity(payload);
+            } else {
+                TS.log().warn("payload was null so ignoring");
+            }
+        } catch (final Exception e) {
+            throw new RuntimeException(e);
+        }
+        return this;
     }
 
 }
