+++ testah-junit/src/main/java/org/testah/driver/http/requests/PutRequestDto.java
@@ -2,6 +2,7 @@ package org.testah.driver.http.requests;
 
 import org.apache.http.HttpEntity;
 import org.apache.http.client.methods.HttpPut;
+import org.testah.TS;
 
 public class PutRequestDto extends AbstractRequestDto {
 
@@ -22,6 +23,20 @@ public class PutRequestDto extends AbstractRequestDto {
     public PutRequestDto(final String uri, final HttpEntity payload) {
         super(new HttpPut(uri), "PUT");
         setPayload(payload);
+    }
+
+    public AbstractRequestDto setPayload(final HttpEntity payload) {
+        try {
+            if (null != payload) {
+                httpEntity = payload;
+                ((HttpPut) httpRequestBase).setEntity(payload);
+            } else {
+                TS.log().warn("payload was null so ignoring");
+            }
+        } catch (final Exception e) {
+            throw new RuntimeException(e);
+        }
+        return this;
     }
 
 }
