+++ testah-junit/src/main/java/org/testah/framework/dto/StepActionDto.java
@@ -174,6 +174,11 @@ public class StepActionDto {
         return this;
     }
 
+    public StepActionDto setException(final Throwable exception) {
+        this.exception = exception;
+        return this;
+    }
+
     public String getMessage1() {
         return message1;
     }
